+++ NEWS.md
@@ -1,10 +1,14 @@
 # santoku (development version)
 
+* `chop_n()` gains a `tail` argument, to deal with a last interval containing
+  less than `n` elements. Set `tail = "merge"` to merge it with
+  the previous interval. This guarantees that all intervals contain at least
+  `n` elements.
 * `chop_equally()` may return fewer than `groups` groups when there are 
   duplicate elements. We now warn when this happens.
 * Bugfix: `chop_n()` could return intervals with fewer than `n` elements 
   when there were duplicate elements. The new algorithm avoids this, but
-  may be slower when there are duplicated elements.
+  may be slower in this case.
 
 # santoku 0.9.1